refactor(dataset): extract Image into vailabel-dataset crate
Phase 1 step 8a. Moves the Image asset + ProjectIdPayload/ImageRangePayload
DTOs into crates/dataset (domain/contracts), impl core::Identifiable, serde
default via vailabel_shared::now_iso. images/model.rs is now a re-export shim;
the orphan-violating local 'impl HasId for Image' is removed. The image
repository/service stay in the binary and operate on the re-exported type.
