+++ apps/studio/src-tauri/Cargo.toml
@@ -12,6 +12,7 @@ members = [
     "crates/plugin",
     "crates/runtime",
     "crates/project",
+    "crates/dataset",
     "crates/annotation",
     "crates/search",
     "crates/training",
@@ -33,6 +34,7 @@ chrono = { version = "0.4", features = ["serde"] }
 vailabel-core = { path = "crates/core" }
 vailabel-shared = { path = "crates/shared" }
 vailabel-project = { path = "crates/project" }
+vailabel-dataset = { path = "crates/dataset" }
 vailabel-plugin = { path = "crates/plugin" }
 vailabel-runtime = { path = "crates/runtime" }
 vailabel-annotation = { path = "crates/annotation" }
@@ -74,6 +76,7 @@ runtime-manager = { path = "crates/runtime-manager" }
 vailabel-core = { workspace = true }
 vailabel-shared = { workspace = true }
 vailabel-project = { workspace = true }
+vailabel-dataset = { workspace = true }
 diesel = { version = "2", features = ["sqlite"] }
 libsqlite3-sys = { version = "0.28", features = ["bundled"] }
 reqwest = { version = "0.12", default-features = false, features = ["blocking", "rustls-tls", "json"] }
