+++ apps/studio/src-tauri/crates/dataset/src/contracts/mod.rs
@@ -1,0 +1,5 @@
+//! Public request/response DTOs for the Dataset module.
+
+pub mod requests;
+
+pub use requests::{ImageRangePayload, ProjectIdPayload};
