+++ apps/studio/src-tauri/crates/dataset/src/domain/mod.rs
@@ -1,0 +1,5 @@
+//! The Dataset domain layer.
+
+pub mod image;
+
+pub use image::Image;
